Refactor dataloader
